+++ Rdmp.Core/CommandExecution/AtomicCommandFactory.cs
@@ -201,10 +201,10 @@ namespace Rdmp.Core.CommandExecution
             if(Is(o,out  AggregateConfiguration ac) && !ac.Catalogue.IsApiCall())
             {
                 yield return new ExecuteCommandCreateNewFilter(_activator, ac) { SuggestedCategory = Add, OverrideCommandName = "New Filter" };
-                yield return new ExecuteCommandCreateNewFilterFromCatalogue(_activator, ac) { SuggestedCategory = Add, OverrideCommandName = "Exisiting Filter" };
+                yield return new ExecuteCommandCreateNewFilterFromCatalogue(_activator, ac) { SuggestedCategory = Add, OverrideCommandName = "Existing Filter" };
 
                 yield return new ExecuteCommandAddNewFilterContainer(_activator,ac) { SuggestedCategory = Add, OverrideCommandName = "New Filter Container" };
-                yield return new ExecuteCommandImportFilterContainerTree(_activator, ac) { SuggestedCategory = Add, OverrideCommandName = "Exisiting Filter Container (copy of)" };
+                yield return new ExecuteCommandImportFilterContainerTree(_activator, ac) { SuggestedCategory = Add, OverrideCommandName = "Existing Filter Container (copy of)" };
 
                 yield return new ExecuteCommandAddParameter(_activator, ac, null, null, null) { SuggestedCategory = Add, OverrideCommandName = "New Catalogue Filter Parameter" };
 
@@ -453,6 +453,14 @@ namespace Rdmp.Core.CommandExecution
 
             if(Is(o, out IFilter filter))
             {
+                if(filter.GetAllParameters().Any())
+                {
+                    yield return new ExecuteCommandSet(_activator,
+                        () => _activator.SelectOne("Select Parameter to change Value for...", filter.GetAllParameters().OfType<IMapsDirectlyToDatabaseTable>().ToArray()),
+                        typeof(ISqlParameter).GetProperty(nameof(ISqlParameter.Value))
+                    ) 
+                    { OverrideCommandName = "Set Parameter Value(s)", Weight = -10 };
+                }
                 yield return new ExecuteCommandViewFilterMatchData(_activator, filter, ViewType.TOP_100);
                 yield return new ExecuteCommandViewFilterMatchData(_activator, filter, ViewType.Aggregate);
             }