+++ CHANGELOG.md
@@ -8,9 +8,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ...
 
+### Added
+
+- "parameter description" and "property name" have been added to the "set value" option for filters - https://github.com/HicServices/RDMP/issues/1034
+- Filter parameter values are now prompted for the user when adding existing filter without known good value sets - https://github.com/HicServices/RDMP/issues/1030
+- "Set Parameter Value(s)" option added to filter menus so you can more easily change the parameter values - https://github.com/HicServices/RDMP/issues/1035
+
 ### Fixed
 
 - Fixed bug preventing example datasets being created from the RDMP UI client because checkbox was disabled
+- "Exisiting" filter typo corrected - https://github.com/HicServices/RDMP/issues/1029
+
+### Changed
+
+- New filters are now highlighted correctly when added to a CIC - https://github.com/HicServices/RDMP/issues/1031
+
 
 ## [7.0.9] - 2022-03-29